+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * Allow additional directories, beyond the default `clj[sc]`, to be correctly formulated by `clojure-expected-ns` via new `defcustom` entitled `clojure-directory-prefixes`
 * Recognize babashka projects (identified by the presence of `bb.edn`).
 
+### Changes
+
+* [#604](https://github.com/clojure-emacs/clojure-mode/pull/604): Add `bb` (babashka) to `interpreter-mode-alist`.
+
 ## 5.13.0 (2021-05-05)
 
 ### New features